backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic-settings==2.5.0

# LlamaIndex core + integrations
llama-index-core==0.12.0
llama-index-llms-openai==0.3.0
llama-index-embeddings-openai==0.3.0
llama-index-vector-stores-supabase==0.3.0

# LlamaParse (LlamaCloud)
llama-cloud>=1.0

# Supabase client
supabase==2.9.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.9
